skribe_api_scraper/requirements.txt
httpx==0.27.0
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.1
pyjwt==2.8.0
rapidfuzz==3.6.1
loguru==0.7.2
streamlit==1.32.1
